# hustlecoin-mix-ui/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HustleCoin Mix API（设计稿反推 v0.1 · 联调基线）
  version: 0.1.0
  description: |
    由 25 块设计画板反推的前后端契约。铁律：
    1) PhaseCode/状态枚举以 /meta/enums 为单一来源，前端绝不推断状态；
    2) 排序键用后端字段 opened_at / pnl；
    3) 子账户行下发 executingAccount + accountKind（hedge_via_master 显主账号）；
    4) 写操作一律 202 受理 + WS position:updates 对账，前端 inflight 防重入；
    5) merged/self 口径必须显式参数，无默认。
servers:
  - url: /api/v1
paths:
...
  /positions:
    get:
      summary: 坑位行（主控台平铺；币种行兼列头 + 账户子行）
      parameters:
        - { name: view, in: query, schema: { type: string, enum: [flat, grouped] }, required: true }
        - { name: sort, in: query, schema: { type: string, enum: [opened_at, pnl] }, required: true }
        - { name: dir, in: query, schema: { type: string, enum: [asc, desc] }, required: true }
        - { name: strategy, in: query, schema: { $ref: '#/components/schemas/StrategyCode' } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PositionRow' } } } }
  /positions/{id}/actions:
    post:
      summary: 行内/右键菜单干预（补单/补对冲/手动还币/强制平仓/移/还…）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, description: 见 CONTEXT_MENUS key }
                accountId: { type: string, description: 子账户行操作时必传 }
                idempotencyKey: { type: string }
      responses:
        '202': { description: 受理（结果走 WS position:updates 对账） }
        '409': { description: 状态机拒绝（如开仓态拆腿闸） }
...
  /report/pnl:
    get:
      summary: 收益序列（柱形图数据源；周/月由前端聚合日粒度）
      parameters:
        - { name: range, in: query, required: true, schema: { enum: [30d, 90d, 180d, all] } }
      responses: { '200': { description: '[{date, funding, spread, fee, net}]' } }
...
  /me/earnings/summary:
    get:
      summary: 今日/本周/累计（合并口径显式）
      parameters: [{ name: view, in: query, required: true, schema: { enum: [merged, self] } }]
      responses: { '200': { description: OK } }
...
  /me/subaccounts:
    get:
      parameters: [{ name: view, in: query, required: true, schema: { enum: [merged, self] } }]
      responses: { '200': { description: OK } }
...
components:
  schemas:
    StrategyCode: { type: string, enum: [S1, S2, S3, S4, S5, S6] }
    PhaseCode:
      type: string
      description: 仲裁契约状态机，唯一来源 /meta/enums；前端直接渲染不推断
      enum: [CANDIDATE, ARBITRATING, ARMED, OPENING, HOLDING, EXITING, SETTLED,
             BORROWABLE, PENDING_BORROW, REPAYING, COOLDOWN_3045, FROZEN,
             QUOTA_CHECK, LENDING, ACCRUING, RECLAIMING,
             EVENT_FEED, EVALUATING, MANUAL_CONFIRM, REVERTING]
    CellValue:
      type: object
      properties: { value: { type: string }, tone: { enum: [up, down, muted, accent, strategy] }, dim: { type: boolean } }
    ParamPair:
      type: object
      properties: { label: { type: string }, value: { type: string }, tone: { type: string } }
    SubRowState:
      type: object
      description: 三态互斥：api_error(红)/borrowing(青)/repay_paused(金,可恢复)/holding/plain
      properties: { kind: { enum: [api_error, borrowing, repay_paused, holding, plain] }, text: { type: string }, resumable: { type: boolean } }
    AccountSubRow:
      type: object
      required: [executingAccount, accountKind, venue, platformType, values, state]
      properties:
        executingAccount: { type: string, description: hedge_via_master 下为平台主账号 }
        accountKind: { enum: [master, sub] }
        venue: { type: string }
        platformType: { enum: [cex, kms_wallet] }
        values: { type: array, items: { $ref: '#/components/schemas/CellValue' }, description: 与主行 columnLabels 同列位对齐 }
        econParams: { type: [array, 'null'], items: { $ref: '#/components/schemas/ParamPair' }, description: 未借到为 null }
        state: { $ref: '#/components/schemas/SubRowState' }
        fundingRateRatio: { type: [string, 'null'] }
        borrowRatePerSec: { type: [string, 'null'], description: 每账户 UID 权重余量不同 }
        banCountdown: { type: [string, 'null'], description: 'M:SS' }
        apiRestricted: { type: boolean }
        apiStatus: { enum: [ok, restricted, healing] }
    PositionRow:
      type: object
      required: [id, symbol, positionCount, strategyCode, phase, phaseLabel, columnLabels, marketParams, ruleScope, subRows]
      properties:
        id: { type: string }
        symbol: { type: string }
        positionCount: { type: integer }
        mark: { enum: [dead, risk] }
        strategyCode: { $ref: '#/components/schemas/StrategyCode' }
        phase: { $ref: '#/components/schemas/PhaseCode' }
        phaseLabel: { type: string }
        columnLabels: { type: array, items: { type: string }, description: 币种行兼列头；混合平铺每行自带本策略标签 }
        marketParams: { type: array, items: { $ref: '#/components/schemas/ParamPair' }, description: 开/平/资/时/限/息 }
        pushStatus: { type: string }
        fundingRateRatio: { type: string }
        singleRuleBrief: { type: string }
        allowRemove: { type: boolean }
        allowRepay: { type: boolean }
        pnl: { type: [number, 'null'] }
        openedAt: { type: [string, 'null'], format: date-time }
        keyDeadlineTs: { type: [integer, 'null'], description: 距今<30min 前端整行渐变 }
        ruleScope: { enum: [template, override] }
        subRows: { type: array, items: { $ref: '#/components/schemas/AccountSubRow' } }
...